refactor(pi-fence): move CommentNode from parse.ts to types.ts
types.ts is the natural home for domain types; parse.ts now only
contains getCommentHash. All import sites updated to reference types.js.
