+++ packages/pi-fence/src/languages/c.test.ts
@@ -13,5 +13,4 @@ describe("c — extractCommentNodes", () => {
     expect(nodes).toHaveLength(1);
     expect(nodes[0].text).toBe("/* ===== section ===== */");
   });
-
 });
