+++ packages/pi-fence/src/languages/javascript.test.ts
@@ -7,5 +7,4 @@ describe("javascript — extractCommentNodes", () => {
     expect(nodes).toHaveLength(1);
     expect(nodes[0].text).toBe("// ---- section ----");
   });
-
 });
