+++ packages/pi-fence/src/types.ts
@@ -1,4 +1,15 @@
-import type { CommentNode } from "./parse.js";
+export interface CommentNode {
+  /** Raw comment text including markers (e.g. `// ---- section ----`). */
+  text: string;
+  /** 1-indexed line number. */
+  startLine: number;
+  /** 0-indexed column. */
+  startCol: number;
+  /** 1-indexed end line (inclusive). */
+  endLine: number;
+  /** 0-indexed end column (exclusive — first column after the comment). */
+  endCol: number;
+}
 
 export interface Finding {
   relativePath: string;
